PAP22-245j, PAP22-251j initial upload and first code
Convention for this files and suggestion: To distinguish the experiments, which are conducted seperate, a 'j' (jona) or 's' (stephan) should be added to the file name.
Thereby the experiments which were conducted together could either be left the way they are, or a 'js' (ordered alphabetically ;)) could be added.